Split MCP HTTP transport tests
